refactor(capabilities): make ui a submodule that owns its kinds
